Move the `Prism` class, add Prism singleton
